+++ sop-example/sop-springmvc/src/main/java/com/gitee/app/config/OpenServiceConfig.java
@@ -18,6 +18,12 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @EnableNacosDiscovery(globalProperties = @NacosProperties(serverAddr = "127.0.0.1:8848"))
 public class OpenServiceConfig extends SpringMvcServiceConfiguration {
+
+
+    public static final String SPRING_APPLICATION_NAME = "spring.application.name";
+    public static final String SERVER_IP = "server.ip";
+    public static final String SERVER_PORT = "server.port";
+
     static {
         ServiceConfig.getInstance().setDefaultVersion("1.0");
     }
@@ -33,12 +39,16 @@ public class OpenServiceConfig extends SpringMvcServiceConfiguration {
     protected void doAfter() {
         super.doAfter();
         try {
+            System.setProperty(SPRING_APPLICATION_NAME, serviceId);
             String ip = NetUtils.localIP();
             namingService.registerInstance(serviceId, ip, port);
+            System.setProperty(SERVER_IP, ip);
+            System.setProperty(SERVER_PORT, String.valueOf(port));
             log.info("注册到nacos, serviceId:{}, ip:{}, port:{}", serviceId, ip, port);
         } catch (NacosException e) {
             log.error("注册nacos失败", e);
             throw new RuntimeException("注册nacos失败", e);
         }
     }
+
 }